+++ app/sprinkles/admin/assets/local/widgets/js/users.js
@@ -29,13 +29,14 @@ function attachUserForm() {
  * Enable/disable password fields when switch is toggled
  */
 function toggleChangePasswordMode(el, userName, changePasswordMode) {
+    var form = el.find("form");
     if (changePasswordMode == 'link') {
         $(".controls-password").find("input[type='password']").prop('disabled', true);
         // Form submits password reset request
-        var form = el.find("form");
-        form
-        .prop('method', 'POST')
-        .prop('action', site.uri.public + '/api/users/u/' + userName + '/password-reset');
+        form.attr({
+            method: 'POST',
+            action: site.uri.public + '/api/users/u/' + userName + '/password-reset'
+        });
 
         var validator = form.validate();
         if (validator) {
@@ -54,9 +55,10 @@ function toggleChangePasswordMode(el, userName, changePasswordMode) {
     } else {
         $(".controls-password").find("input[type='password']").prop('disabled', false);
         // Form submits direct password update
-        el.find("form")
-        .prop('method', 'PUT')
-        .prop('action', site.uri.public + '/api/users/u/' + userName + '/password');
+        form.attr({
+            method: 'PUT',
+            action: site.uri.public + '/api/users/u/' + userName + '/password'
+        });
     }
 }