+++ app/sprinkles/core/assets/local/core/js/uf-form.js
@@ -53,16 +53,15 @@
         this.element = element[0];
         this.$element = $(this.element);
         var lateDefaults = {
-            reqParams: {
-                type    : this.$element.attr('method'),
-                url     : this.$element.attr('action')
-            },
             encType  : (typeof this.$element.attr('enctype') !== 'undefined') ? this.$element.attr('enctype') : '',
             msgTarget: this.$element.find('.js-form-alerts:first')
         };
         this.settings = $.extend(true, {}, defaults, lateDefaults, options);
-        this._defaults = defaults;
+        this._defaults = $.extend(true, {}, defaults, lateDefaults);
         this._name = pluginName;
+
+        // Detect changes to element attributes
+        this.$element.attrchange({ callback: function (event) { this.element = event.target; }.bind(this) });
 
         // Expose settings for 'onkeyup' until a better more event driven apporch is adopted.
         var settings = this.settings;
@@ -87,19 +86,25 @@
                     submitButton.html(this.settings.submittingText);
                 }
 
+                // Get basic request parameters.
+                var reqParams = {
+                    type: this.$element.attr('method'),
+                    url: this.$element.attr('action')
+                };
+
                 // Get the form encoding type from the users HTML, and chose an encoding form.
                 if (this.settings.encType.toLowerCase() === "multipart/form-data" ) {
-                    this.settings.reqParams.data = this._multipartData(form);
+                    reqParams.data = this._multipartData(form);
                     // add additional params to fix jquery errors
-                    this.settings.reqParams.cache = false;
-                    this.settings.reqParams.contentType = false;
-                    this.settings.reqParams.processData = false;
+                    reqParams.cache = false;
+                    reqParams.contentType = false;
+                    reqParams.processData = false;
                 } else {
-                    this.settings.reqParams.data = this._urlencodeData(form);
+                    reqParams.data = this._urlencodeData(form);
                 }
 
                 // Submit the form via AJAX
-                $.ajax(this.settings.reqParams).then(
+                $.ajax(reqParams).then(
                     // Submission successful
                     $.proxy(function(data, textStatus, jqXHR) {
                         // Restore button text and re-enable submit button
@@ -186,8 +191,9 @@
          */
         _urlencodeData: function(form) {
             // Serialize and post to the backend script in ajax mode
+            var serializedData;
             if (this.settings.binaryCheckboxes) {
-                var serializedData = form.find(':input').not(':checkbox').serialize();
+                serializedData = form.find(':input').not(':checkbox').serialize();
                 // Get unchecked checkbox values, set them to 0
                 form.find('input[type=checkbox]:enabled').each(function() {
                     if ($(this).is(':checked')) {
@@ -198,7 +204,7 @@
                 });
             }
             else {
-                var serializedData = form.find(':input').serialize();
+                serializedData = form.find(':input').serialize();
             }
 
             return serializedData;
@@ -216,7 +222,7 @@
                 var checkboxes = form.find('input[type=checkbox]:enabled');
                 // Feature detection. Several browsers don't support `set`
                 if (typeof formData.set !== 'function') {
-                    this.settings.msgTarget.ufAlerts("push", "danger", "Your browser is missing a required feature. This form will still attempt to submit, but if it fails, you'll need to use Chrome for desktop or FireFox for desktop.")
+                    this.settings.msgTarget.ufAlerts("push", "danger", "Your browser is missing a required feature. This form will still attempt to submit, but if it fails, you'll need to use Chrome for desktop or FireFox for desktop.");
                 }
                 else {
                     checkboxes.each(function() {
@@ -262,20 +268,3 @@
         }
     };
 })(jQuery, window, document);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-