+++ app/sprinkles/core/locale/en_US/messages.php
@@ -47,6 +47,7 @@ return [
 
     "NAME"  => "Name",
     "NAVIGATION" => "Navigation",
+    "NO_RESULTS" => "Sorry, we've got nothing here.",
 
     "PAGINATION" => [
         "GOTO" => "Jump to Page",