+++ CHANGELOG.md
@@ -14,11 +14,15 @@
 - Automatically load Sprinkler service providers (see #636)
 - Get rid of "implicit loading" for core Sprinkle - core is now just an ordinary Sprinkle like any other.
 
+## v.4.0.22-Alpha
+- Fix issue where 'Change User Password' popup form couldn't handle specifying a new password.
+- Display message when there are no results in `ufTable`
+
 ## v.4.0.21-Alpha
 - Implement reflow and column selector for tables (#670)
 - Overhauled ufAlerts, improving efficiency, reliability, and fixed a discovered edge case that caused `render` to never complete. (part of #646)
 - ufAlerts will only auto-scroll when outside the viewport (even if only partially). Can be overriden with `scrollWhenVisible: true`. (#714)
-- Rebased ufCollection, ufForm, ufModal, and ufTable with new jQuery plugin template. (part of #646)
+- Rebased ufCollection, and ufForm with new jQuery plugin template. (part of #646)
 - Misc UI update
 - Added Twig blocks
 - Fix issue with duplicate query logs when using multiple databases